added supporting tests for only the complex circle group (exp, compose, is_identity)
added scalar_multiplication_operation.jl to improve readability

todo:
 -> add support for testing references as input
 -> further functions

many of the problems regarding the implementation of the scalar case root in a lack of functionality in ManifoldsBase.copy_to! and ManifoldsBase.allocate_result() for manifolds with an empty representation_size therefore a scalar representation. for a cleaner and more general implementation those methods need to be generalized